requirement.txt
gradio
transformers
torch
mysql-connector-python
pandas
